renames spanningtree module to stp
This commit renames the spanningtree api module to stp to align with naming
conventions for other modules.  The spanningtree api module will still work
but generate a deprecation warning when used.  Unit and system tests
updated as well to reflect the module change
